bin/sppsh: add account option
Sppsh is only accept an account name on host. It means that you have to
use ssh command if you login with other account.

This update is for adding account option to login any of accounts. Here
is an example of login foo@bar.
'sppsh -a foo bar' or 'sppsh -a foo 0' if bar has VM ID 0.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>