bin/wakeup: enable to change portmask from config
Portmask is assigned from '-p' option or default value if the option is
not given. This scheme is defined in 'primary.py' and user cannot change
it from 'wakeup.py' or config. However, all of parameters for SPP
primary and secondaries should be changed from 'wakeup.py' or config
which is most used from users.

This update is to change assigning portmask from 'primary.py' to config.
'wakeup.py' loads portmask from config and gives it to 'primary.py'.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>